requirements.txt

streamlit==1.36.0
pandas==2.2.2
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.34.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
